fix-修复获取限时分红树时int和double转换问题

+++ lib/provider/money_group.dart
@@ -245,7 +245,7 @@ class MoneyGroup with ChangeNotifier {
           grade: Tree.MAX_LEVEL,
           type: TreeType.Type_TimeLimited_Bonus,
           duration: invite_award.duration,
-          amount: invite_award.amount,
+          amount: invite_award.amount.toDouble(),
           showCountDown: true,
           treeId: invite_award.tree_id,
           timePlantedLimitedBonusTree: DateTime.now().millisecondsSinceEpoch,
